Not available files support

+++ app/src/main/java/com/smartcloud/web/HTMLCreator.java
@@ -5,6 +5,7 @@ import com.smartcloud.database.ServerDatabase;
 import com.smartcloud.holder.FileHolder;
 
 import java.lang.reflect.InvocationTargetException;
+import java.util.ArrayList;
 import java.util.List;
 
 public class HTMLCreator {
@@ -54,35 +55,49 @@ public class HTMLCreator {
 
     public static String cloudFilesHTML() {
         StringBuilder sb = new StringBuilder("<h2>Files</h2><ul>");
-        for (FileHolder fileHolder : ServerDatabase.instance.selectFile(true)) {
-            sb.append("<li>");
-            sb.append("<a href='/fileId=");
-            sb.append(fileHolder.getId());
-            sb.append("' download='");
-            sb.append(fileHolder.getName());
-            sb.append("'>");
-            sb.append(fileHolder.getName());
-            sb.append(" (");
-            sb.append(fileHolder.getSizeReadable());
-            sb.append(") ");
-            sb.append("</a>");
-            sb.append("<a href='/deleteFileId=");
-            sb.append(fileHolder.getId());
-            sb.append("' style='color:red'>-");
-            sb.append("</a>");
-            sb.append("</li>");
-        }
-        for (FileHolder fileHolder : ServerDatabase.instance.selectFile(false)) {
-            sb.append("<li>");
-            sb.append(fileHolder.getName());
-            sb.append(" (");
-            sb.append(fileHolder.getSizeReadable());
-            sb.append(") ");
-            sb.append("<a href='/deleteFileId=");
-            sb.append(fileHolder.getId());
-            sb.append("' style='color:red'>-");
-            sb.append("</a>");
-            sb.append("</li>");
+        for (FileHolder fileHolder : ServerDatabase.instance.selectFile()) {
+            if (!fileHolder.isAvailable()) {
+                sb.append("<li style='text-decoration:line-through;'>");
+                sb.append(fileHolder.getName());
+                sb.append(" (");
+                sb.append(fileHolder.getSizeReadable());
+                sb.append(") ");
+                sb.append("<a href='/deleteFileId=");
+                sb.append(fileHolder.getId());
+                sb.append("' style='color:red;'>-");
+                sb.append("</a>");
+                sb.append("</li>");
+                continue;
+            }
+            if (fileHolder.isActive()) {
+                sb.append("<li>");
+                sb.append("<a href='/fileId=");
+                sb.append(fileHolder.getId());
+                sb.append("' download='");
+                sb.append(fileHolder.getName());
+                sb.append("'>");
+                sb.append(fileHolder.getName());
+                sb.append(" (");
+                sb.append(fileHolder.getSizeReadable());
+                sb.append(") ");
+                sb.append("</a>");
+                sb.append("<a href='/deleteFileId=");
+                sb.append(fileHolder.getId());
+                sb.append("' style='color:red'>-");
+                sb.append("</a>");
+                sb.append("</li>");
+            } else {
+                sb.append("<li>");
+                sb.append(fileHolder.getName());
+                sb.append(" (");
+                sb.append(fileHolder.getSizeReadable());
+                sb.append(") ");
+                sb.append("<a href='/deleteFileId=");
+                sb.append(fileHolder.getId());
+                sb.append("' style='color:red'>-");
+                sb.append("</a>");
+                sb.append("</li>");
+            }
         }
         sb.append("</ul><br>");
         return sb.toString();
@@ -100,15 +115,31 @@ public class HTMLCreator {
             sb.append("<h3>Machines Inactive</h3>").append(tableDomainHTML(new String[]{"Id", "Machine Role", "Total Space", "Last Contact"}, list,
                     new String[]{"getId", "getMachineRole", "getTotalSpaceReadable", "getLastContactFormat"}));
         }
-        list = ServerDatabase.instance.selectFile(true);
-        if (!list.isEmpty()) {
-            sb.append("<h3>Files Active</h3>").append(tableDomainHTML(new String[]{"Id", "Name", "Size"}, list, new String[]{"getId", "getName", "getSizeReadable"}));
+        list = ServerDatabase.instance.selectFile();
+        List<FileHolder> filesActive = new ArrayList<>();
+        List<FileHolder> filesInactive = new ArrayList<>();
+        List<FileHolder> filesUnavailable = new ArrayList<>();
+        for (FileHolder fileHolder : ServerDatabase.instance.selectFile()) {
+            if (!fileHolder.isAvailable()) {
+                filesUnavailable.add(fileHolder);
+                continue;
+            }
+            if (fileHolder.isActive()) {
+                filesActive.add(fileHolder);
+            } else {
+                filesInactive.add(fileHolder);
+            }
         }
-        list = ServerDatabase.instance.selectFile(false);
-        if (!list.isEmpty()) {
-            sb.append("<h3>Files Inactive</h3>").append(tableDomainHTML(new String[]{"Id", "Name", "Size"}, list, new String[]{"getId", "getName", "getSizeReadable"}));
+        if (!filesActive.isEmpty()) {
+            sb.append("<h3>Files Active</h3>").append(tableDomainHTML(new String[]{"Id", "Name", "Size"}, filesActive, new String[]{"getId", "getName", "getSizeReadable"}));
         }
-        list = ServerDatabase.instance.selectSegment(null);
+        if (!filesInactive.isEmpty()) {
+            sb.append("<h3>Files Inactive</h3>").append(tableDomainHTML(new String[]{"Id", "Name", "Size"}, filesInactive, new String[]{"getId", "getName", "getSizeReadable"}));
+        }
+        if (!filesUnavailable.isEmpty()) {
+            sb.append("<h3>Files Unavailable</h3>").append(tableDomainHTML(new String[]{"Id", "Name", "Size"}, filesUnavailable, new String[]{"getId", "getName", "getSizeReadable"}));
+        }
+        list = ServerDatabase.instance.selectSegment();
         if (!list.isEmpty()) {
             sb.append("<h3>Segments</h3>").append(tableDomainHTML(new String[]{"Id", "File Id", "Machine Id", "Byte From", "Byte To", "Size"}, list,
                     new String[]{"getId", "getFileId", "getMachineId", "getByteFrom", "getByteTo", "getSizeReadable"}));
